+++ apt-pkg/versionmatch.cc
@@ -143,8 +143,8 @@ bool pkgVersionMatch::MatchVer(const char *A,string B,bool Prefix)
    const char *Ae = Ab + strlen(A);
    
    // Strings are not a compatible size.
-   if (((unsigned)(Ae - Ab) != B.length() && Prefix == false) ||
-       (unsigned)(Ae - Ab) < B.length())
+   if (((size_t)(Ae - Ab) != B.length() && Prefix == false) ||
+       (size_t)(Ae - Ab) < B.length())
       return false;
    
    // Match (leading?)
